CRAS: client - use a pipe instead of a barrier.
Using the pipe allows the main thread to notice that the client thread
has closed the other end.  The client thread will close its end of the
pipe when it exits.  Using the pthread barrier led to a window where the
main thread could start waiting for a command to complete after the
client thread had exited.

BUG=chromium:128084
TEST=Build and run chrome/cras for z600 and alex.  On alex do "stop cras
&& stop ui", observe that chrome doesn't crash with a sigabrt.

Change-Id: I54e57118d62b171a6b840389b4737cb76c999811
Signed-off-by: Dylan Reid <dgreid@chromium.org>