CRAS: iodev_list - don't route away from plugged device.
When sorting the device list, don't move streams that are not on the
default device.  Streams move to non-default devices when a jack is
plugged and when a device has a plugged jack it should remain the
default until it is unplugged.  Fixes an issue with routing away from
HDMI/DP devices at startup.

BUG=chromium-os:20619
TEST=updated UT, boot stumpy with DP plugged, head audio from DP device.

Signed-off-by: Dylan Reid <dgreid@chromium.org>
(cherry picked from commit abb898da72badf589c27960d493eca93bb64972d)

Change-Id: Icad0c84ef33badd1774588d74f9c28d18b61c12b