CRAS: system_state - remove de-init code from init.
The clean up code in init was only there for the unit test, since the
system state lives for the life of the audio server.  Move that code to
a function that is only called from the unit test.

BUG=chromium-os:32278
TEST=update system_state unittest.

Change-Id: I111fbbb9ce8aabcb0ae321a71864ec850435be0c
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/27606
Reviewed-by: Olof Johansson <olofj@chromium.org>
(cherry picked from commit 16418c77d91b99b482fa36756bc32ced776aa933)
Reviewed-on: https://gerrit.chromium.org/gerrit/28749