CRAS: iodev - Save jack plugged state and time.
When a jack is plugged, save the state (plugged/unplugged) and the time
of the last plug in the iodev.  This state and timestamp will be used to
decide the most recently plugged device when more than one device could
be plugged.

BUG=chromium-os:32650
TEST=updated iodev_unittest.

Change-Id: I28eba0434a9e9e3bfb5f61d5ac4c2b710c7ccaeb
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/27712
Reviewed-by: Olof Johansson <olofj@chromium.org>
(cherry picked from commit 8284c1b74b0b4ee012aa7e1566940abf81ae2ef5)
Reviewed-on: https://gerrit.chromium.org/gerrit/28754