CRAS: iodev_list - Route to recently plugged when detach current.
When the currently playing device gets a jack removal event, route to
the device with the next most recent plug time if there is any.  If
there isn't another plugged device, route to the device most recently
added to the system.  For example if HDMI is plugged then headphones are
plugged, when the headphones are unplugged the audio should route back
to HDMI, not to internal speakers.  Previously only the most recently
attached device got audio when a jack was unplugged.

BUG=chromium-os:32650
TEST=updated unit tests and try these reproduction steps:
1. play youtube video
2. plug in Audio capable monitor or tv with DP to HDMI adapter
3. hear audio on TV
4. Plug in headphones to lumpy/stumpy
5. hear audio on headphones, not on TV.
6. unplug headphones.
7. Hear audio from TV, not from built in speakers.

Change-Id: I2519f9474d5134a198be842f28a1baafa89f7538
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/27714
(cherry picked from commit 31942965ffd73c76fd946c5e28a204ecc198ed58)
Reviewed-on: https://gerrit.chromium.org/gerrit/28757