CRAS: Route only based on card type plugged state.
This patch does several things to normalize the choosing of a default
device.  This decision is made whenever a device is added or removed or
when a jack event arrives.

Remove the sorting of device lists at startup, the priorities of the
device are sufficient to determine the correct place to route audio.

Use this same logic to pick the default device when a device is
attached/detached, and when jack events happen.

BUG=chromium-os:32730
TEST=updated unit tests, Boot lumpy with HP, USB, and DP connected, plug
and unplug the devices and listen to audio routing to the most recently
plugged of HP/USB or HDMI if neither of those two are attached.

Change-Id: I3388f336003823bd89e4a3c704bde05fde46d041
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/27946
(cherry picked from commit 19b14f1bbc18b05185898408d227bc4c176dc7c6)
Reviewed-on: https://gerrit.chromium.org/gerrit/28761