CRAS: alsa plug - override default pcm_delay function.
Provide an implementation to get the delay of the next sample to be
captured or played.  The default implementation is backwards in lib alsa
and returns negative numbers.  This neglects some server-side delay but
it should be more than good enough for now.  The server-side delay can
be added later, but I want to keep this change small for applying to
R21.

For playback the delay is the amount the user has filled the buffer past
the hardware pointer.  For capture it is the opposite, how much data is
left to be read is the delay.

BUG=chromium-os:33021
TEST=run program to print latency for capture and playback, make sure
they are sane.  Test the AEC in GTalk that relies on accurate delay
times.

Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/28620
Reviewed-by: Olof Johansson <olofj@chromium.org>
(cherry picked from commit bf77ce9bf2f0ea82485da8e6548873de090285b0)

Change-Id: I002ef0100083fac21269c7ec32bc1d53efbea84c
Reviewed-on: https://gerrit.chromium.org/gerrit/29495
Reviewed-by: Dylan Reid <dgreid@chromium.org>
Tested-by: Dylan Reid <dgreid@chromium.org>