CRAS: alsa_card - Create a UCM instance for each card.
If there is a configuration available, each alsa card will now get a ucm
instance.  This instance is passed to each iodev in the card.  This UCM config
will be used to do microphone switching for ASoC.  The controls that need to be
switched to enable the external mic are not easy to determine.

BUG=chrome-os-partner:10849
TEST=updated unit tests.

Change-Id: Ibfbebdb5cb27f8f49f1baa68b475ddac53c3f597
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/30296