CRAS: alsa_io - use UCM to enable mic if available.
If there is a UCM configuration for the card, attempt to enable the
headset device when there in a Mic Jack event.  This assumes that there
is only a Mic input jack(It is all that is searched for), and that that
input wants to turn on the Headset ucm device.  In the future there
needs to be a better way to map a jack to a ucm input.

BUG=chrome-os-partner:10849
TEST=Run on Daisy with UCM config loaded, plug/unplug headset and check
that the routing switches with amixer.

Change-Id: Ieaae8ecac174fc6181f9fc2943741771fe0e01f6
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/30298