CRAS: gpio jack - Check that jacks match the card.
GPIO jacks were being added to all cards in the system.  Luckily the
name of the card is pre-pended to the name of the jack, so it is easy to
check if the card name is passed in to the jack creation routine.

Remember to close the gpio fd when destroying the jack list.

BUG=chromium-os:34375
TEST=update UT, plug headphones while playing to USB, check system
doesn't hang.

Change-Id: Ibca79e382fa78e986a91813ad1c57ad7834a3938
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/33141
(cherry picked from commit 71e47f74597f17ecbb95f7fa929c6b2b6353a06d)
Reviewed-on: https://gerrit.chromium.org/gerrit/34356