CRAS: audio thread: remove stream if iodev open fails.
When adding a stream to a thread, if the device fails to open, the
procedure is repeated on the next best device.  In this case the
stream wasn't removed from the newly created thread and when the
thread was destroyed an extra message was sent to the client telling
it that the stream was removed.

BUG=chromium:190557
TEST=cause iodev open failure (by removing device) and check that the
stream moves smoothly to the next device.

Change-Id: Ic9e8c994a7e7c01937ff7b922b41eca4a8727e4f
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/47825
Reviewed-by: Chih-Chung Chang <chihchung@chromium.org>
(cherry picked from commit 8a9dee38e0096f85f32682c9eab389e45c80ca85)
Reviewed-on: https://gerrit.chromium.org/gerrit/48332