CRAS: Update name of headphone node.
The name of the headphone node was changed to be more specific for the device
selection UI.  This broke the check if the headphone was plugged.  Update the
string to match the headphone node name on stumpy, the only system that uses
this API.

BUG=chromium:234397
TEST=set short suspend time and watch powerd log for message about not
suspending because of the headphone jack.

Signed-off-by: Dylan Reid <dgreid@chromium.org>
Change-Id: Idab2ca4231f7acfbd4ac8fe3c68ae7558802660a
Reviewed-on: https://gerrit.chromium.org/gerrit/49092
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
(cherry picked from commit f6777fcb8ea7c6c6a931f23616d1e696aec2e2fc)
Reviewed-on: https://gerrit.chromium.org/gerrit/49591