CRAS: audio_thread - don't adjust sleep for extra USB buffer
This code was belt-and-suspenders anyways and it was getting confused
by offsetting the sleep time for the stream.  The sleep time should
only be based on the stream level, not the hw level.

Sleeping for too short a time was causing repeated audio samples.

BUG=chromium:180560
TEST=start hangout from gchat with usb headset attached.

Change-Id: Idbc8d34f9b42b72f43f64e21b69793d3db0b6239
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/49631
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
(cherry picked from commit 0048b7d7c3a863e9f2aa11abe72cd4fd88625596)
Reviewed-on: https://gerrit.chromium.org/gerrit/50077