CHERRY-PICK: CRAS: audio_thread - fetch based on adjusted amount.
The decision to fetch audio from a client during playback should be
based on the adjusted level.  When there is a min_buffer_level set, the
callback needs to be made earlier so that that level can be maintained.
Worse, the sleep level was already accounting for the adjustment, leading
to always sleeping for 0 samples for devices that have adjusted levels.

BUG=chrome-os-partner:25220
TEST=Play youtube on panther, check cpu usage.  Added unit test to check
this is done correctly.

Change-Id: I4ea2dbf77ff103e35a2d1cc6d52def6a9dfb036e
Original-Change-Id: Ibf440a0cfcfb9e5853009dd7e7f8ba7a46a32da2
Original-Change-Id: I02c82876e330cd42993d48e15bf48bda76c2bc22
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/184450
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
Reviewed-by: Chih-Chung Chang <chihchung@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/184698
Commit-Queue: Mohammed Habibulla <moch@chromium.org>
Tested-by: Mohammed Habibulla <moch@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/189613
Reviewed-by: Mohammed Habibulla <moch@google.com>
Commit-Queue: Mohammed Habibulla <moch@google.com>
Tested-by: Mohammed Habibulla <moch@google.com>