nyan ucm: Decrease analog mic gain.
This was set for a total gain of over 40dB, drop it to 10dB which is
more reasonable for a headset mic that is likely close to speaker.
The AGC will still move this gain in hangouts, but 10dB is a better
default for various flash apps that use it.

BUG=chrome-os-pertner:28402
TEST=record audio with online-voice-recorder.com on Big.

Change-Id: I9d66bff58ae899a4752234c41cf60f72a13fd41a
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/197560
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
(cherry picked from commit 527feb79a0f40b09848961de96d0c070eff48ed6)
Reviewed-on: https://chromium-review.googlesource.com/200106