CRAS: audio_thread - Handle dev error while appending stream
If there is a device error while appending a stream to the device,
return an error from append_stream_to_dev and handle it in
append_stream.  This will ensure that the stream lands on the fallback
device and that the fallback device is initialized.  Before it could
be dropped and the fallback device could be left uninitialized.

BUG=chromium:425286
TEST=select device, open it with aplay -Dhw to generate an error and
play a new stream.

Change-Id: I51c485973934de9ef7b629ab438af9a73fd53fb8
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/224476
(cherry picked from commit 6b3e2865a03f831d1e350f8003f268634f799ec9)
Reviewed-on: https://chromium-review.googlesource.com/224987