CRAS: Don't limit frames by offset in capture.
The write point returned from cras_shm_get_writeable_frames needs to
have enough space for the offset and the number of remaining frames.
This many frames will be written starting at offset. For example: an
offset of 60 in to a 100 deep buffer, the pointer will be returned to
the first frame (frame zero), Then 50 frames could be written
beginning at the offset of 60, because the shm only limited based on
its write point which is still at the beginning, so there is room for
50 frames.

cras_audio_area_dopy already handles limiting the total number based
on the offset.

BUG=chromium:427675
TEST=record from two devices at once on Blaze.  Jabra and built-in.

Change-Id: I416fc83d460fe1941c5ae08793fa59611d6e72a9
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/225703
Reviewed-by: Chinyue Chen <chinyue@chromium.org>
Reviewed-by: Hsinyu Chao <hychao@chromium.org>
(cherry picked from commit 5d5525afc084e1d00faab2fdbce85ed458b2fc53)
Reviewed-on: https://chromium-review.googlesource.com/226104