UCM config: Samus: adjust default ADC1 capture volume after TLV change
When we fixed the volume TLV, the resolution of
"ADC1 Capture Volume" was halved, but its default
value in the ucm config was not changed:

UPSTREAM: ASoC: rt5677: make volume TLV closer to reality
https://chromium-review.googlesource.com/233550

A value of 63 used to mean -17.625+63*0.375 = 6dB; but after
the patch, it means -17.625+(63*2+1)*0.375 = 30dB now. With
a "STO1 ADC Boost Volume" of 24dB, the overall gain is too
big and causing distortion.

This patch changes the default value to 31 which means
-17.625+(31*2+1)*0.375 = 6dB.

BUG=chrome-os-partner:35688
TEST=arecord from DMICs and recording is clean

Change-Id: Ifef9af7388432985ca82b582593ee1eb12020560
Signed-off-by: Ben Zhang <benzh@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/242741
Reviewed-by: Dylan Reid <dgreid@chromium.org>
(cherry picked from commit d5a5f93fd3728e3b4deceeca6ade005ed36b1d85)
Signed-off-by: Ben Zhang <benzh@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/243366