ucm: ryu: adjust default ADC1 capture volume after TLV change
When we fixed the volume TLV, the resolution of
"ADC1 Capture Volume" was halved, but its default
value in the ucm config was not changed:

UPSTREAM: ASoC: rt5677: make volume TLV closer to reality
https://chromium-review.googlesource.com/233550

A value of 63 used to mean -17.625+63*0.375 = 6dB; but after
the patch, it means -17.625+(63*2+1)*0.375 = 30dB now. With
a "STO1 ADC Boost Volume" of 24dB, the overall gain is too
big and causing distortion.

This patch changes the default value to 31 which means
-17.625+(31*2+1)*0.375 = 6dB.

Based on Samus change https://chromium-review.googlesource.com/#/c/242741/

BUG=None
TEST=P3 mic recording

Change-Id: Ib9e22be0f8402f4a2049e972d0e8b1ebceb82aac
Reviewed-on: https://chromium-review.googlesource.com/247221
Reviewed-by: David Riley <davidriley@chromium.org>
Tested-by: Anatol Pomazau <anatol@google.com>