CRAS: Ignore master volume if extra specified
Change the semantics of ExtraMainVolume.  Instead of adding to the
master control, make it replace the master control.  The master control
on Samus is fairly useless compared to the DAC1 control in the codec.
Since Samus is the only user of this option, just change it.

BUG=chrome-os-partner:34332
TEST=listen to hot for teacher and adjust volume.

Change-Id: Ia95ee3b9df9310516c100030b8689d5631a316fe
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/248160
Reviewed-by: Ben Zhang <benzh@chromium.org>
Reviewed-by: Anatol Pomazau <anatol@google.com>
(cherry picked from commit 9798eb637c7f7a971052ae02703629e0942a369c)
Reviewed-on: https://chromium-review.googlesource.com/249560