CRAS: ucm - Add MinBufferLevel variable
This variable will be used to specify the minimum buffer level for a
device.  The Rockchip devices don't implement a residue for their DMA,
so the internal output needs to be treated like a USB device where the
hw_level can't be trusted and should be padded.

BUG=chrome-os-partner:38754

Change-Id: I4a875fd504abaaf9b7404f062a20448f88eb09e4
Signed-off-by: Dylan Reid <dgreid@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/264024
(cherry picked from commit 1200364bd86be0ca80fe2cd5615f696580b62e82)
Reviewed-on: https://chromium-review.googlesource.com/264141