CRAS: iodev - Set rate in device format before allocating DSP
DSP needs a sampling rate to be initialized.
The rate in stream might not be the same as the rate of the device.
Since DSP is applied right before we write samples to device, we should
use the rate of device to initialize DSP.
In cras_iodev_set_format, we should query the actual rate from device
and set the rate in ext_format and format.

BUG=chrome-os-partner:47898
TEST=Plays a 44100 rate stream.Check the sampling rate of DSP is 48000 on samus.

Signed-off-by: Cheng-Yi Chiang <cychiang@chromium.org>
Change-Id: Id578024776949575d8d7b908137b67c6f846e488
Previous-Reviewed-on: https://chromium-review.googlesource.com/313851
(cherry picked from commit e2f511bd4ee92a4f96b51d15137f4cfbf83103fb)
Reviewed-on: https://chromium-review.googlesource.com/314540