CRAS: alsa_io - Change name of the node created by jack
In legacy node creation flow, jack name will be used as node name.
In this case the node fails to get the property set in UCM.
For example, mic node on Daisy is named "Daisy-I2S Mic Jack".
It can not get the properties set in "Mic" section in UCM.
This CL fixes this by setting the node name to common name "Mic".

Another solution is to change the section name in the UCM but it is
better to use the common name as section name in the UCM.

BUG=chromium:761765,chromium:747450
TEST=unittest
TEST=check node name on peach_pit is common name.
Change-Id: Idce617f51a711c457c20a88fc40b02e5bcbfc6d4
Previous-Reviewed-on: https://chromium-review.googlesource.com/650147
(cherry picked from commit d2d23dd1479164c04f9168bc7ca76da03d5d86eb)
Reviewed-on: https://chromium-review.googlesource.com/683916
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>
Commit-Queue: Cheng-Yi Chiang <cychiang@chromium.org>
Tested-by: Cheng-Yi Chiang <cychiang@chromium.org>