CRAS: cras_iodev_list - Enable ucm setting of active node when resetting
When we reset audio device, we should disable and enable ucm setting of
active node. We only disable but not enable it in current version. The
change is for fixing this bug.
Adding it in cras_iodev_list_enable_dev function instead of enable_device
because it will not affect much functions and it should be ok to enable
ucm twice.

BUG=b:116066435
TEST=Reseting the audio device works fine on eve

Change-Id: I4b5c50539b5872346977bb8f2c95d5bc6605e00e
Previous-Reviewed-on: https://chromium-review.googlesource.com/1234514
(cherry picked from commit 48632e6ee3610e9274f1f71bb242f57d61074ce2)
Reviewed-on: https://chromium-review.googlesource.com/1249405
Reviewed-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Commit-Queue: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Tested-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>