CRAS: bt_device - Skip connected callback for non-audio devices
When a bluetooth device is connected, the server will go into the
bt_device_conn_watch_cb function. However, it doesn't consider the
situation that the connected device is not an audio device. It causes
an audio device to be removed when other non-audio devices connect.

To fix it, we can check the profile at the beginning of the device
connected callback. If it is not an audio device, skip the callback.

Notice that it is only a quick fix. The better way is to avoid non-audio
devices from getting into this callback. We will implement it on ToT to
avoid a large-scale change.

BUG=b:139021513, chromium:755242
TEST=The BT headphones work fine when pairing other BT devices.

Change-Id: I152cf3fc592bce15820cf36a68ff264a312b5aa6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/1761763
Reviewed-by: En-Shuo Hsu <enshuo@google.com>
Commit-Queue: En-Shuo Hsu <enshuo@google.com>
Commit-Queue: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Tested-by: En-Shuo Hsu <enshuo@google.com>
(cherry picked from commit 67d2a74ff8bf381c4805a724a76cdd28ee4cee01)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/1763519
Reviewed-by: Chih-Yang Hsia <paulhsia@chromium.org>
Commit-Queue: Chih-Yang Hsia <paulhsia@chromium.org>
Tested-by: Chih-Yang Hsia <paulhsia@chromium.org>