ucm-config: Fixup for Pixel USB-C earbuds on R81
On R81 we are missing two fixes for system gain and plug detection.
These flags are obsolete on R82 and above, but are needed in R81:

JackType "always"
MaxSoftwareGain
DefaultNodeGain

This CL fixes the previous CL:

14875183 ucm: Pixel USB-C earbuds

so that Pixel USB-C earbuds can be used.

BUG=b:151794046
TEST=On R81, apply this patch. Check Pixel USB-C earbuds is detected,
and the gain is large enough.

Change-Id: I7171cd076bb4b35900261d9194ee079e1ca7dbd1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/2124270
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>
Commit-Queue: Cheng-Yi Chiang <cychiang@chromium.org>
Tested-by: Cheng-Yi Chiang <cychiang@chromium.org>