ucm-config: Add config for PCP-USB
The user of PCP-USB specifically requested that the gain to be tuned at
30% for professional usage.
Set that accordingly and fix that gain.

User can still apply software gain through API or through system UI.

Note that on R81, we need to set an extra value:
JackType "Always" for detection to work.

BUG=chromium:1071786
TEST=check PCP-USB has mixer control "Mic Capture Volume" fixed at 19.

Change-Id: I128ba039dd5ee194548bded9f7de79df2f48077c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/2153419
Reviewed-by: Yu-Hsuan Hsu <yuhsuan@chromium.org>
Reviewed-by: Ben Zhang <benzh@chromium.org>
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>
Commit-Queue: Cheng-Yi Chiang <cychiang@chromium.org>
Tested-by: Cheng-Yi Chiang <cychiang@chromium.org>