CRAS: device_monitor - Close erroneous device in main thread
If an output device hit I/O error in audio thread, there was
no explicit code path to close it. So if application keeps
the audio output stream alive, the output device would keep the
same buffer playing forever.

This change adds below handling logic:
When EIO is received, check if it's within 10 seconds from the
last EIO error. If not, then reset this device(close and reopen).
Otherwise treat this as a non-stop error and close this device
to avoid audio looping.

BUG=b:170100252
TEST=Use code to fake EIO error after 5 seconds of playback,
verify the device receives a reset, and 5 more seconds later
it gets closed gracefully.

Change-Id: Ic90a5a59c70d7f87bc7e52ce243235c15df5a13d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/2543707
Reviewed-by: Rehan Ghori <rehang@chromium.org>
Reviewed-by: Joe Tessler <jrt@chromium.org>
Tested-by: Rehan Ghori <rehang@chromium.org>
Commit-Queue: Hsinyu Chao <hychao@chromium.org>