ucm: enable hotword modifier on device open
With SOF restoring controls on resume, when WoV is not enabled we see a
resume time regression because the kernel is copying the hotword model
to the firmware. The solution is to only set the model when we are using
the device.

This patch is one part of the solution in that we trigger
the disable sequence on the modifier when close the hotword device. The
second part will be to set a null model on the modifier's disable
sequence.

The only side effect will be non SOF boards is that the model will be
set each time right before device opens.

BUG=b:175897062
TEST=check resume time and verify WoV works

Cq-Depend: chromium:2612115
Change-Id: Icd3a0103b5b5e267992ee065200f77b628208741
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/2601896
Reviewed-by: Ben Zhang <benzh@chromium.org>
Tested-by: Curtis Malainey <cujomalainey@chromium.org>
Commit-Queue: Curtis Malainey <cujomalainey@chromium.org>
(cherry picked from commit 7dd7322b3ee0e3e1d814c78bb924328627b453ad)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/adhd/+/2616747
Reviewed-by: Curtis Malainey <cujomalainey@chromium.org>